Fixing the overall project structure to be up-to-standard for express/javascript applications.
