+++ public/js/app.js
@@ -1,0 +1,4 @@
+// Browser / client-side configDotenv, not related to node.js
+document.addEventListener('DOMContentLoaded', () => {
+    JobController.init();
+});
